pyproject.toml
[project]
name = "crud-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faker>=40.12.0",
    "fastapi>=0.135.3",
    "gradio>=6.11.0",
    "openai>=2.30.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
